Added files to the repo
Step 1: Make sure you have Python 3.10+ installed on your machine.
Step 2: run the .bat file which will create a venv folder and activate the virtual environment and then the python file will install the needed modules and the script would run

Recommended: Using the snipping tool, go to youtube.com click on any video and when you see an ad, pause the video, using the snipping tool, snip the skip-Ad button and save it inside the folder as "skip_ad_button.png". This is a one time activity. Incase you change the resolution of your monitor, or connect to a monitor with a different resolution, you need to repeat this process.